Refactored package name to match maven namespace, added publishing plan
