--- lib/src/commonMain/kotlin/com/shambu/compose/scrollbar/foundation/ScrollbarMeasurements.kt
@@ -1,9 +1,0 @@
-package com.shambu.compose.scrollbar.foundation
-
-import androidx.compose.ui.geometry.Rect
-
-data class ScrollbarMeasurements(
-    val barBounds: Rect,
-    val indicatorBounds: Rect,
-    val alpha: Float,
-)
