+++ lib/src/commonMain/kotlin/io/github/abhijithshambu/scrollbar/foundation/ScrollbarMeasurements.kt
@@ -1,0 +1,9 @@
+package io.github.abhijithshambu.scrollbar.foundation
+
+import androidx.compose.ui.geometry.Rect
+
+data class ScrollbarMeasurements(
+    val barBounds: Rect,
+    val indicatorBounds: Rect,
+    val alpha: Float,
+)
