--- sample/composeApp/src/androidMain/kotlin/com/shambu/compose/scrollbar/sample/MainActivity.kt
@@ -1,18 +1,0 @@
-package com.shambu.compose.scrollbar.sample
-
-import MainView
-import android.os.Bundle
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.appcompat.app.AppCompatActivity
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        enableEdgeToEdge()
-        super.onCreate(savedInstanceState)
-
-        setContent {
-            MainView()
-        }
-    }
-}
