+++ sample/composeApp/src/androidMain/kotlin/io/github/abhijithshambu/scrollbar/sample/MainActivity.kt
@@ -1,0 +1,18 @@
+package io.github.abhijithshambu.scrollbar.sample
+
+import MainView
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        enableEdgeToEdge()
+        super.onCreate(savedInstanceState)
+
+        setContent {
+            MainView()
+        }
+    }
+}
