--- sample/composeApp/src/commonMain/kotlin/com/shambu/compose/scrollbar/sample/theme/Colors.kt
@@ -1,22 +1,0 @@
-package com.shambu.compose.scrollbar.sample.theme
-
-import androidx.compose.ui.graphics.Color
-
-val md_theme_light_primary = Color(0xFF8800EC)
-val md_theme_light_onPrimary = Color(0xFFFFFFFF)
-val md_theme_light_secondary = Color(0xFF5555A9)
-val md_theme_light_onSecondary = Color(0xFFFFFFFF)
-val md_theme_light_error = Color(0xFFBA1A1A)
-val md_theme_light_onError = Color(0xFFFFFFFF)
-val md_theme_light_background = Color(0xFFFFFBFF)
-val md_theme_light_onBackground = Color(0xFF1D1B1E)
-val md_theme_light_surface = Color(0xFFEFE7ED)
-val md_theme_light_onSurface = Color(0xFF1D1B1E)
-
-val md_theme_dark_error = Color(0xFFFFB4AB)
-val md_theme_dark_onError = Color(0xFF690005)
-
-val violet_700 = Color(0xFF090919)
-val violet_300 = Color(0xFF181626)
-val violet_vibrant = Color(0xFF940CFF)
-val white_light = Color(0xBBFFFFFF)
