+++ sample/composeApp/src/commonMain/kotlin/io/github/abhijithshambu/scrollbar/sample/data/Models.kt
@@ -1,0 +1,16 @@
+package io.github.abhijithshambu.scrollbar.sample.data
+
+data class AlbumModel(
+    val title: String,
+    val description: String,
+    val songs: List<SongModel>,
+    val imageUrl: String,
+    val style: String,
+)
+
+data class SongModel(
+    val title: String,
+    val description: String,
+    val imageUrl: String,
+    val album: String,
+)
